fixing the circular error
