--- backend/learnwell_backend/user_functionality/__init__.py
@@ -1,1 +1,0 @@
-from models import topic, user_entity, study_session, study_plan, break_activity
